Strong instantiation of a LogManagerBase for JNI
The wrappers were providing a weak definition of the default LogManager instance, but this failed for customers who were Java-only, and did not appear to work even for a unit test that provided a LogManager definition. Since we have Java-only customers, it seems reasonable to supply a solution that supports them.